\name{simCTMC-package}
\alias{simCTMC-package}
\alias{simCTMC}
\docType{package}
\title{
\packageTitle{simCTMC}
}
\description{
\packageDescription{simCTMC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{simCTMC}
\packageIndices{simCTMC}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{simCTMC}

Maintainer: \packageMaintainer{simCTMC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
