\name{simCTMC-package}
\alias{simCTMC-package}
\alias{simCTMC}
\docType{package}
\title{
\packageTitle{simCTMC}
}
\description{
\packageDescription{simCTMC}
}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{simCTMC}
\packageIndices{simCTMC}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}

\author{
\packageAuthor{simCTMC}

Maintainer: \packageMaintainer{simCTMC}
}

